Code review: Fix straight flush detection miss in cardsvalue.cpp when 6+ cards share same suit (old code only checked 3 fixed contiguous windows j1..j1+4 for j1<3, missing valid straight flushes when extra same-suit cards broke the rank chain - e.g. 6 hearts with K,10,9,8,7,6 missed straight flush 10-9-8-7-6 because King broke the consecutive check at j1=0 and j1=2 hit the spade at position 6; replace with suit-group scanning that finds extent of each suit group then checks all possible 5-card windows within it, also fixes wheel straight flush A-5-4-3-2 to search within suit group instead of fixed positions), fix guilog.cpp integer overflow in sqlite3_get_table where (nRow+1)*nCol computed as int can overflow for large query results causing undersized malloc buffer overrun (use size_t arithmetic), fix ConfigFile::writeBuffer() not flushing/closing file after writing (add stream.flush() and file.close() matching all other write paths in updateConfig), remove duplicate FlopValues entries in arraydata.cpp (hand code 50207 appeared 3 times lines 652-654 and hand code 53201 appeared 2 times lines 746-747, linear scan always matched first entry making duplicates dead code bloating the array)

+++ src/gui/qt/gametable/log/guilog.cpp
@@ -108,7 +108,7 @@ extern "C" int sqlite3_get_table(sqlite3 *pDb, const char *zSql, char ***pazResu
 	}
 
 	int nRow = rows.size();
-	int total = (nRow + 1) * nCol;
+	size_t total = static_cast<size_t>(nRow + 1) * static_cast<size_t>(nCol);
 
 	struct FreeDeleter {
 		void operator()(char** p) const { if(p) { for(char **ptr = p; *ptr; ++ptr) free(*ptr); free(p); } }
